src/functions/requirements.txt
awscrt
aws-iot-device-sdk-python
paho-mqtt